Refactor widget routes
